+++ app/src/main/java/com/patloew/oeffitracker/ui/trip/list/TripItem.kt
@@ -58,7 +58,7 @@ import java.time.LocalDate
 @Composable
 fun TripItem(
     trip: Trip?,
-    onDelete: (id: Int) -> Unit,
+    onDelete: (id: Long) -> Unit,
     onDuplicateForToday: (Trip) -> Unit
 ) {
     if (trip != null) {
